Question 1.1.5 Extent to which the included deployment guide is well written and clearly articulates the steps necessary to deploy the system.
Response 4: Required  In your samples it would be clearer to specify the endpoint being called first, and then the parameters, not the parameters first, especially when there are multiple parameters preceding the endpoint. It's not clear to see the parameters when we don't know yet what endpoint is being called.